Now supporting MSTATUS.SD bit.
Now supporting MSTATUS.VS field: If VS is off and an attempt to execute a
vector instruction is made then we get an illegal instruction. If VS is
not off and a vector instruction is attempted, then VS is marked dirty.

Now clearing VSTART after each vector instruction unless an exception
is encoutered.

Updated default mask of SSTATUS.