Make it safe to call PerfApi::drainStore() on stores that had exceptions
I think encouraging the perf model to call drainStore() more often is
good from a standpoint of triggering the cleanup code at the end of this
function that ensures we remove entries from the packet and store maps,
so we don't leak memory.

Therefore, I would like to make it safe to call drainStore() even if a
store encountered an exception and did not complete (or perhaps only
partially completed, when we support vectors).  This has the effect of
making drainStore() more consistent with the current behavior of
retire(), which the perf model calls even on instructions that
experience exceptions and do not retire in hardware in the traditional
sense.

In the case of scalar stores, I've only observed them either complete
entirely or not at all, so here I check if dsize_ is zero indicating the
store did not complete, and in that case it should just drain nothing
and only clean up the packet and store maps.